Change bench input to integer, print benching time.

+++ src/lib.rs
@@ -29,9 +29,9 @@ pub(crate) struct Args {
     /// File path for the part 2 memory trace
     part2_file: Option<String>,
 
-    #[structopt(long, default_value = "3.0")]
-    /// Benchmarking period to measure run time of parts.
-    bench_time: f64,
+    #[structopt(long, default_value = "3")]
+    /// Benchmarking period in seconds to measure run time of parts.
+    bench_time: u32,
 }
 
 pub struct IsExample;
